requirements.txt
Flask==3.0.0
nltk==3.8.1
scikit-learn==1.3.2
google-generativeai==0.7.1
gunicorn==22.0.0